requirements.txt
argon2-cffi==20.1.0
bcrypt==3.1.7
cffi==1.14.1
Django==2.2
django-bootstrap-form==3.4
django-bootstrap3==14.1.0
django-cors-headers==3.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
Pillow==7.2.0
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
